pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "difference calculator"
authors = ["Alexandr Myshkin <sir.myshkin@gmail.com>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^3.0.0"
flake8 = "^5.0.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"